Add API routes for AI and Autumn integrations
- Introduced a new `api-routes.ts` file to handle API endpoints for AI and Autumn functionalities.
- Implemented middleware for database connection, authentication, and session management.
- Added error handling for API requests.
- Updated `main.ts` to utilize the new API routes, streamlining the server's route management.
