+++ apps/server/src/routes/ai.ts
@@ -1,6 +1,8 @@
 import { CallService } from '../services/call-service/call-service';
+import { getAgentByName, routeAgentRequest } from 'agents';
 import { ZeroMCP } from '../services/mcp-service/mcp';
 import { env } from 'cloudflare:workers';
+import type { ZeroAgent } from '../main';
 import twilio from 'twilio';
 import { Hono } from 'hono';
 
@@ -24,6 +26,20 @@ aiRouter.mount(
   },
   { replaceRequest: false },
 );
+
+aiRouter.post('/chat/:userId', async (c) => {
+  const userId = c.req.param('userId');
+
+  if (!userId) {
+    return c.text('Missing userId', 400);
+  }
+
+  // Retrieve a stub to the specific ZeroAgent instance
+  const stub = await getAgentByName<Env, ZeroAgent>(env.ZERO_AGENT, userId);
+
+  // Forward the incoming request body straight to the Agent
+  return await stub.fetch(c.req.raw);
+});
 
 aiRouter.post('/voice', async (c) => {
   const formData = await c.req.formData();
